requirements.txt
lightning==2.2.1
einops==0.7.0
sentencepiece==0.2.0
beartype==0.18.2
typing==3.7.4.3
torcheval==0.0.7
transformers[torch]==4.40.1
#pytorch-ignite==0.5.0.post2
#datasets==2.18.0